gh#35: Fix template processing to work with ESI
ESI elements are in the form <esi:include src="...">; note the
namespace. Namespaces aren't a valid part of HTML apparently, so
lxml.html was stripping them out, breaking ESI entirely.

This update switches our template handling (although not all parts
of the conversion process I think) to using lxml.etree. Its XPath
syntax is more awkward, in that you apparently *always* have to
specify a namespace for the element you're searching for (that
was pretty confusing) but as far as I could make out that was the
way to get our namespaces left in the markup when the converted
tree was written out.

Also I suppose if we like robust, valid markup (and who doesn't?)
it's nice to use a proper XML parser rather than an HTML one which
might be more liberal?